[FLINK-17611] Refactor to improve testability
